refactor: Refactor cmd_table_traversal_utils
1. Separate traversal functions into new file.
2. Rename move_to_end_of_pipeline() to move_past_pipeline bc the old name was not correct.
3. Merge the subshell traversal functions bc only going past the subshell is needed.
