+++ source/backend/executor/handle_control_op.c
@@ -17,7 +17,7 @@
 #include "clean.h"
 
 bool	should_execute_next_pipeline(int type, int exit_code)
-{	
+{
 	if (type == C_AND && exit_code == 0)
 		return (true);
 	else if (type == C_OR && exit_code != 0)
@@ -34,8 +34,7 @@ void	handle_and_or_op(t_shell *shell, t_list_d **cmd_table_node)
 	cmd_table = (*cmd_table_node)->content;
 	*cmd_table_node = (*cmd_table_node)->next;
 	if (!should_execute_next_pipeline(cmd_table->type, shell->exit_code))
-		// move_past_pipeline(cmd_table_node);
-		move_to_end_of_pipeline(cmd_table_node);
+		move_past_pipeline(cmd_table_node);
 }
 
 void    handle_control_op(t_shell *shell, t_list_d **cmd_table_node)
